Tried to make Visual Studio happy

+++ include/domain.hpp
@@ -43,7 +43,11 @@ struct peanokey_hash : public thrust::unary_function<point_t<T>, peanokey>
   __host__ __device__
   peanokey operator()(point_t<T> p) const
   {
-    return peano_hilbert_key(p.x, p.y, p.z, 23);
+    return peano_hilbert_key(
+		static_cast<float>(p.x), 
+		static_cast<float>(p.y), 
+		static_cast<float>(p.z), 
+		23);
   }
 };
